personal_finance_coach/requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
python-dotenv==1.0.0
Authlib==1.3.1
requests>=2.31.0
Flask-Session==0.4.0
marshmallow>=3.20.0

# Development/Linting tools
flake8>=7.0.0
black>=24.0.0
pylint>=3.0.0
autopep8>=2.0.0
isort>=5.0.0